+++ src/homekit/climate/active.ts
@@ -12,10 +12,14 @@ export abstract class ActiveClimateAccessory<C extends ActiveClimateConfig = Act
 
     this.setup(HKCharacteristicKey.Active, dependency.Characteristic.Active.INACTIVE,
       'topicGetActive',
-      this.bindOnUpdateNumericBoolean(HKCharacteristicKey.Active, 'valueStateActive', strings.active.active, strings.active.notActive, true), true,
+      this.bindOnUpdateNumericBoolean(HKCharacteristicKey.Active, 'valueStateActive', strings.active.active, strings.active.notActive, false, (value) => {
+        this.onSetActive(value === 1);
+      }), true,
       'topicSetActive',
       this.bindOnSetBoolean(HKCharacteristicKey.Active, 'topicSetActive', 'valueStateActive', 'valueStateInactive', dependency.Characteristic.Active.ACTIVE,
-        strings.active.activeSet, strings.active.inactiveSet),
+        strings.active.activeSet, strings.active.inactiveSet, (value) => {
+          this.onSetActive(value);
+        }),
     );
 
     this.setup(HKCharacteristicKey.LockPhysicalControls, dependency.Characteristic.LockPhysicalControls.CONTROL_LOCK_DISABLED,
@@ -51,5 +55,20 @@ export abstract class ActiveClimateAccessory<C extends ActiveClimateConfig = Act
       this.bindOnSetBoolean(HKCharacteristicKey.SwingMode, 'topicSetSwingMode', 'valueSwingEnabled', 'valueSwingDisabled',
         dependency.Characteristic.SwingMode.SWING_ENABLED, strings.active.swingEnabledFuture, strings.active.swingDisabledFuture),
     );
+  }
+
+  private onSetActive(active: boolean) {
+
+    if (!active) {
+      return;
+    }
+
+    this.startTimeout(() => {
+      this.onSetBoolean(HKCharacteristicKey.Active, false, 'topicSetActive', 'valueStateActive', 'valueStateInactive',
+        this.Characteristic.Active.ACTIVE, strings.active.activeSet, strings.active.inactiveSet,
+        () => {
+          this.onSetActive(false);
+        });
+    });
   }
 }